feat(backend): add backend Docker Hub build script
This commit adds a new file `backend-docker-hub.sh` to the address book
deployment in the Kubernetes backend directory. The script `backend-docker-hub.sh`
contains the command to build the Docker image for the address book backend,
tagged as `address-book-backend:1.0.0`. The build command uses the Dockerfile
located in the `address-book-backend` directory.

The purpose of this script is to streamline the process of building and
tagging the Docker image for the backend application, making it easier to
deploy the address book backend to Docker Hub.